ARROW-4502: [C#] Add support for zero-copy reads
Update to the latest Google FlatBuffers code to support Spans/Memory.

Add a constructor for ArrowStreamReader that takes a ReadOnlyMemory<byte>.
Add a synchronous ReadNextRecordBatch() method.

Since we are now enabling Spans with FlatBuffers, we need to change the way we write to streams in the ArrowStreamWriter to use Memory<byte> instead of byte[]. This API is in netcoreapp2.1, but not in netstandard, so cross compile for netcoreapp2.1 and add a shim for netstandard.